Car_Price/requirements.txt
pandas==1.4.4
streamlit-lottie==0.0.3
jsonschema==4.16.0
Pillow==9.2.0
numpy==1.23.3
pickle-mixin==1.0.2
streamlit==1.13.0
scikit-learn==1.1.2
protobuf==3.20.1